# docs/openapi-password-reset-admin0.yaml
openapi: 3.0.3
info:
  title: METRIX — Reset de contraseña de licencia (Admin 0)
  version: 1.0.0-draft
  description: |
    Recuperación de la contraseña del ADMIN de una instancia METRIX.
    Admin 0 aprueba o rechaza. La liga es de un solo uso (24 h).
    Verificar en /swagger-ui/index.html (perfil local).
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
tags:
  - name: Autenticación
  - name: Plataforma
paths:
  /api/v1/auth/password-reset/request:
    post:
      tags: [Autenticación]
      summary: Solicitar reset (olvidé mi contraseña)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequestBody'
      responses:
        '204': { description: Aceptado. Siempre, si el body es válido. }
        '400': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /api/v1/auth/password-reset/validate:
    get:
      tags: [Autenticación]
      summary: Validar token de la liga
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetValidateResponse'
        '422': { $ref: '#/components/responses/Error' }
  /api/v1/auth/password-reset/confirm:
    post:
      tags: [Autenticación]
      summary: Consumir la liga y guardar la nueva contraseña
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
      responses:
        '204': { description: Contraseña actualizada }
        '400': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /api/v1/platform/password-resets:
    get:
      tags: [Plataforma]
      summary: Listar solicitudes de reset
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PasswordResetStatus'
        - name: instanceId
          in: query
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PasswordResetRequestResponse'
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/platform/password-resets/{id}/approve:
    post:
      tags: [Plataforma]
      summary: Aprobar y enviar liga
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetRequestResponse'
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Error' }
  /api/v1/platform/password-resets/{id}/reject:
    post:
      tags: [Plataforma]
      summary: Rechazar solicitud
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRejectRequest'
      responses:
        '204': { description: Rechazada }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/platform/instances/{id}/password-reset:
    post:
      tags: [Plataforma]
      summary: Admin 0 dispara reset del ADMIN de esa licencia
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetRequestResponse'
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RequestId:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
    PasswordResetStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, CONSUMED, EXPIRED]
    PasswordResetRequestBody:
      type: object
      required: [codigoEmpresa, numeroUsuario]
      properties:
        codigoEmpresa: { type: string, example: TACOS-A3F2 }
        numeroUsuario: { type: string, example: ADMIN001 }
    PasswordResetValidateResponse:
      type: object
      properties:
        valid: { type: boolean }
        codigoEmpresa: { type: string }
        empresaNombre: { type: string }
        numeroUsuario: { type: string }
        expiresAt: { type: string, format: date-time }
    PasswordResetConfirmRequest:
      type: object
      required: [token, newPassword, confirmPassword]
      properties:
        token: { type: string, minLength: 32 }
        newPassword: { type: string, minLength: 6 }
        confirmPassword: { type: string }
    PasswordResetRejectRequest:
      type: object
      properties:
        reason: { type: string, maxLength: 280 }
    PasswordResetRequestResponse:
      type: object
      properties:
        id: { type: string }
        instanceId: { type: string }
        codigoEmpresa: { type: string }
        empresaNombre: { type: string }
        numeroUsuario: { type: string }
        adminNombre: { type: string }
        destinationEmailMasked: { type: string, example: a***@tacos.test }
        status: { $ref: '#/components/schemas/PasswordResetStatus' }
        requestedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time, nullable: true }
        resolvedAt: { type: string, format: date-time, nullable: true }
        resolvedBy: { type: string, nullable: true }
        consumedAt: { type: string, format: date-time, nullable: true }
        rejectReason: { type: string, nullable: true }
        resetUrl:
          type: string
          format: uri
          description: Solo en approve / initiate. Nunca en el GET de listado.
        emailSent: { type: boolean }
    ApiError:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        details:
          type: object
          additionalProperties: { type: string }
  responses:
    Error:
      description: Error de negocio o validación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: No es Admin 0
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'